+++ django/applications/catmaid/static/js/CATMAID.js
@@ -18,6 +18,57 @@
       return ui;
     },
   });
+
+  // Configuration of message position
+  var messagePosition = 'tr';
+  Object.defineProperty(CATMAID, 'messagePosition', {
+    get: function() { return messagePosition; },
+    set: function(newValue) {
+      var allowedValues = ['tl', 'tr', 'bl', 'br', 'tc', 'bc'];
+      if (-1 === allowedValues.indexOf(newValue)) {
+        throw new CATMAID.ValueError('Please use one of these values: ' +
+            allowedValues.join(','));
+      }
+      messagePosition = newValue;
+    }
+  });
+
+  /**
+   * Convenience function to show a growl message
+   */
+  CATMAID.msg = function(title, message, options)
+  {
+    var settings = {
+      title: title,
+      message: message,
+      duration: 3000,
+      size: 'large',
+      location: messagePosition,
+      style: undefined // Gray background by default, alternatives are:
+                       // 'error' = red, 'warning' = yellow, 'notice' = green
+    };
+
+    // If an alert style wasn't provided, guess from the alert title
+    if (!options || !options.style) {
+      if (title.match(/error/i)) settings.style = 'error';
+      else if (title.match(/warn|beware/i)) settings.style = 'warning';
+      else if (title.match(/done|success/i)) settings.style = 'notice';
+    }
+
+    $.extend(settings, options);
+    $.growl(settings);
+  };
+
+  /**
+   * Convenience function to show a growl info message.
+   */
+  CATMAID.info = CATMAID.msg.bind(window, "Information");
+
+  /**
+   * Convenience function to show a growl warning message.
+   */
+  CATMAID.warn = CATMAID.msg.bind(window, "Warning");
+
 })(CATMAID);
 
 
@@ -213,21 +264,3 @@ CATMAID.error = function(msg, detail)
 {
   new CATMAID.ErrorDialog(msg, detail).show();
 };
-
-/**
- * Convenience function to show a growl message
- */
-CATMAID.msg = function(title, msg)
-{
-  growlAlert(title, msg);
-};
-
-/**
- * Convenience function to show a growl info message.
- */
-CATMAID.info = CATMAID.msg.bind(window, "Information");
-
-/**
- * Convenience function to show a growl warning message.
- */
-CATMAID.warn = CATMAID.msg.bind(window, "Warning");