+++ django/applications/catmaid/static/js/widgets/graph_plot.js
@@ -3,7 +3,6 @@
 /* global
   CircuitGraphAnalysis,
   fetchSkeletons,
-  growlAlert,
   InstanceRegistry,
   NeuronNameService,
   OptionsDialog,
@@ -232,7 +231,7 @@ CircuitGraphPlot.prototype.append = function(models) {
 
   if (skids.length < 2) {
     this.clearGUI();
-    growlAlert('Need more than one', 'Need at least two neurons!');
+    CATMAID.msg('Need more than one', 'Need at least two neurons!');
     return;
   }
 
@@ -462,7 +461,7 @@ CircuitGraphPlot.prototype.getVectors = function() {
 
   var f = (function(select) {
     var index = select.selectedIndex;
-    if (index < this.vectors.length) {
+    if (this.vectors && index < this.vectors.length) {
       return this.vectors[index][1];
     } else if ('a' === select.value[0]) {
       if (!this.anatomy) {
@@ -669,7 +668,7 @@ CircuitGraphPlot.prototype.loadAnatomy = function(callback) {
       },
       function(skid) {
         // Failed to load
-        growlAlert("ERROR", "Skeleton #" + skid + " failed to load.");
+        CATMAID.msg("ERROR", "Skeleton #" + skid + " failed to load.");
         measurements[skid] = null;
       },
       (function() {